requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
requests-html
faiss-cpu
numpy
openai
python-dotenv